contour: keep the frame scale in float
83624fbe compared the render-target-derived scale against
effectiveDevicePixelRatio() in qreal, then handed the result to
composeItemToClip(), which takes a float -- a narrowing conversion, and
clang-tidy runs with -warnings-as-errors. deviceToLogicalScale() already
returns float; convert the Qt value once and compare in float throughout.

+++ src/contour/display/TerminalRenderNode.cpp
@@ -92,13 +92,13 @@ void TerminalRenderNode::prepare()
     // which is a single value and so cannot pair two moments. Kept as a cross-check rather than the
     // primary source because the render-target-derived ratio is what makes a layered or offscreen target
     // come out right, and it agrees with this whenever the snapshot is current.
-    auto const dpr = [&]() -> qreal {
+    auto const dpr = [&]() -> float {
         auto const derived = deviceToLogicalScale(rt->pixelSize(), windowLogicalSize);
         auto const* const window = display->window();
         if (window == nullptr)
             return derived;
-        auto const effective = window->effectiveDevicePixelRatio();
-        if (effective > 0.0 && std::abs(derived - effective) > 0.001)
+        auto const effective = static_cast<float>(window->effectiveDevicePixelRatio());
+        if (effective > 0.0f && std::abs(derived - effective) > 0.001f)
             return effective;
         return derived;
     }();
